Removes BaseRow class and converts CharSet variable to static
